adjusted Profiles screen
