+++ tailwind.config.js
@@ -18,6 +18,11 @@ module.exports = {
     './app/screens/budgets/BudgetsScreen.{js,jsx,ts,tsx}',
     './app/screens/leaderboards/LeaderBoardsScreen.{js,jsx,ts,tsx}',
     './app/screens/profile/ProfileScreen.{js,jsx,ts,tsx}',
+    './app/screens/profile/components/SignOutButton.{js,jsx,ts,tsx}',
+    './app/screens/profile/components/Profile.{js,jsx,ts,tsx}',
+    './app/screens/profile/components/Friends.{js,jsx,ts,tsx}',
+    './app/screens/profile/components/Settings.{js,jsx,ts,tsx}',
+    './app/screens/profile/components/Extensions.{js,jsx,ts,tsx}',
     './app/screens/budgets/components/BudgetModal.{js,jsx,ts,tsx}',
     './app/screens/budgets/components/AddCategoryModal.{js,jsx,ts,tsx}',
     './app/screens/budgets/components/ChooseIconPopup.{js,jsx,ts,tsx}',
